+++ src/dinoex_jobs.h
@@ -38,8 +38,8 @@ __attribute__ ((format(printf, 2, 3)))
 #endif
 privmsg_chan(const channel_t *ch, const char *format, ...);
 
-void writeserver_privmsg(writeserver_type_e delay, const char *nick, const char *message, int len);
-void writeserver_notice(writeserver_type_e delay, const char *nick, const char *message, int len);
+void writeserver_privmsg(writeserver_type_e delay, const char *nick, const char *message, size_t len);
+void writeserver_notice(writeserver_type_e delay, const char *nick, const char *message, size_t len);
 void clean_send_buffers(void);
 void sendannounce(void);
